+++ js/modules/util.js
@@ -1,0 +1,14 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+
+const getId = () => {
+  let currentId = 0;
+  return () => ++currentId;
+};
+
+export {getRandomInteger, getId};
